+++ tools/scan6.c
@@ -2237,7 +2237,7 @@ int main(int argc, char **argv){
 
 							if(ulhtype == IPPROTO_ICMPV6){
 								if( idata.type == DLT_EN10MB && !(idata.flags & IFACE_LOOPBACK) && pkt_icmp6->icmp6_type == ND_NEIGHBOR_SOLICIT){
-									if( (pkt_end - (unsigned char *) pkt_ns) < sizeof(struct nd_neighbor_solicit))
+									if( (pkt_end - (unsigned char *) pkt_ns) < (signed long int) sizeof(struct nd_neighbor_solicit))
 										continue;
 
 									/* 
@@ -2732,7 +2732,7 @@ puts("After to pcap_next_ex()");
 
 					if(pkt_ipv6->ip6_nxt == IPPROTO_ICMPV6){
 						if( idata.type == DLT_EN10MB && !(idata.flags & IFACE_LOOPBACK) && pkt_icmp6->icmp6_type == ND_NEIGHBOR_SOLICIT){
-							if( (pkt_end - (unsigned char *) pkt_ns) < sizeof(struct nd_neighbor_solicit))
+							if( (pkt_end - (unsigned char *) pkt_ns) < (signed long int) sizeof(struct nd_neighbor_solicit))
 								continue;
 
 							/* 
@@ -2762,7 +2762,7 @@ puts("After to send_neighbor_advert()");
 							if(!is_ip6_in_scan_list(&scan_list, &(pkt_ipv6->ip6_src)))
 								continue;
 
-							if( (pkt_end - (unsigned char *) pkt_icmp6) < sizeof(struct icmp6_hdr))
+							if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) sizeof(struct icmp6_hdr))
 								continue;
 
 							if(valid_icmp6_response_remote(&idata, &scan_list, probetype, pkthdr, pktdata, buffer)){
@@ -5024,7 +5024,7 @@ int multi_scan_local(pcap_t *pfd, struct iface_data *idata, struct in6_addr *src
 
 			if(pkt_ipv6->ip6_nxt == IPPROTO_ICMPV6){
 				if(pkt_icmp6->icmp6_type == ND_NEIGHBOR_SOLICIT){
-					if( (pkt_end - (unsigned char *) pkt_ns) < sizeof(struct nd_neighbor_solicit))
+					if( (pkt_end - (unsigned char *) pkt_ns) < (signed long int) sizeof(struct nd_neighbor_solicit))
 						continue;
 
 					if(is_eq_in6_addr(&(pkt_ns->nd_ns_target), srcaddr) || \
@@ -5036,7 +5036,7 @@ int multi_scan_local(pcap_t *pfd, struct iface_data *idata, struct in6_addr *src
 					}
 				}
 				else if( (pkt_icmp6->icmp6_type == ICMP6_ECHO_REPLY) || (pkt_icmp6->icmp6_type == ICMP6_PARAM_PROB)){
-					if( (pkt_end - (unsigned char *) pkt_icmp6) < sizeof(struct icmp6_hdr))
+					if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) sizeof(struct icmp6_hdr))
 						continue;
 
 					/*
@@ -5346,7 +5346,7 @@ int host_scan_local(pcap_t *pfd, struct iface_data *idata, struct in6_addr *srca
 				if(pkt_icmp6->icmp6_type == ND_NEIGHBOR_SOLICIT){
 					pkt_ns= (struct nd_neighbor_solicit *) pkt_icmp6;
 
-					if( (pkt_end - (unsigned char *) pkt_ns) < sizeof(struct nd_neighbor_solicit))
+					if( (pkt_end - (unsigned char *) pkt_ns) < (signed long int) sizeof(struct nd_neighbor_solicit))
 						continue;
 
 					if(is_eq_in6_addr(&(pkt_ns->nd_ns_target), &(idata->ip6_local)) || \
@@ -5359,7 +5359,7 @@ int host_scan_local(pcap_t *pfd, struct iface_data *idata, struct in6_addr *srca
 				}
 				else if( (pkt_icmp6->icmp6_type == ICMP6_ECHO_REPLY) || (pkt_icmp6->icmp6_type == ICMP6_PARAM_PROB)){
 
-					if( (pkt_end - (unsigned char *) pkt_icmp6) < sizeof(struct icmp6_hdr))
+					if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) sizeof(struct icmp6_hdr))
 						continue;
 
 					if(valid_icmp6_response(idata, type, pkthdr, pktdata, buffer)){
@@ -5626,7 +5626,7 @@ int valid_icmp6_response(struct iface_data *idata, unsigned char type, struct pc
 			   Discard the packet if it is not of the minimum size to contain an ICMPv6 
 			   header and the payload we included in the ICMPv6 Echo Request
 			 */
-			if( (pkt_end - (unsigned char *) pkt_icmp6) < (sizeof(struct icmp6_hdr) + \
+			if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) (sizeof(struct icmp6_hdr) + \
 									ICMPV6_ECHO_PAYLOAD_SIZE) )
 				return 0;
 
@@ -5645,7 +5645,7 @@ int valid_icmp6_response(struct iface_data *idata, unsigned char type, struct pc
 			   Discard the packet if it is not of the minimum size to contain an ICMPv6 
 			   header and the embedded payload
 			 */
-			if( (pkt_end - (unsigned char *) pkt_icmp6) < (sizeof(struct icmp6_hdr) + \
+			if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) (sizeof(struct icmp6_hdr) + \
 						+ sizeof(struct ip6_hdr) + MIN_DST_OPT_HDR_SIZE + sizeof(struct icmp6_hdr) + \
 						  ICMPV6_ECHO_PAYLOAD_SIZE) )
 				return 0;
@@ -5732,7 +5732,7 @@ int valid_icmp6_response_remote(struct iface_data *idata, struct scan_list *scan
 			   Discard the packet if it is not of the minimum size to contain an ICMPv6 
 			   header and the payload we included in the ICMPv6 Echo Request
 			 */
-			if( (pkt_end - (unsigned char *) pkt_icmp6) < (sizeof(struct icmp6_hdr) + \
+			if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) (sizeof(struct icmp6_hdr) + \
 									ICMPV6_ECHO_PAYLOAD_SIZE) )
 				return 0;
 
@@ -5749,7 +5749,7 @@ int valid_icmp6_response_remote(struct iface_data *idata, struct scan_list *scan
 			   Discard the packet if it is not of the minimum size to contain an ICMPv6 
 			   header and the empedded payload
 			 */
-			if( (pkt_end - (unsigned char *) pkt_icmp6) < (sizeof(struct icmp6_hdr) + \
+			if( (pkt_end - (unsigned char *) pkt_icmp6) < (signed long int) (sizeof(struct icmp6_hdr) + \
 						+ sizeof(struct ip6_hdr) + MIN_DST_OPT_HDR_SIZE + sizeof(struct icmp6_hdr) + \
 						  ICMPV6_ECHO_PAYLOAD_SIZE) )
 				return 0;
